Added comments. Changed i2c interface to use uint16_t for sizes. Added new example for soft i2c slave.

+++ include/interface/i2c_interface.h
@@ -27,7 +27,7 @@ public:
     // parameter specifies if a I2C STOP condition
     // is generated at the end.
     virtual int16_t i2cRead (uint16_t addr,
-                             uint8_t *rxbuf, uint8_t len,
+                             uint8_t *rxbuf, uint16_t len,
                              bool sendStop = true) = 0;
 
     // Write len bytes from txbuf to the I2C interface.
@@ -35,7 +35,7 @@ public:
     // negative value in case of errors. The last parameter
     // specifies if a I2C STOP condition is generated at the end.
     virtual int16_t i2cWrite(uint16_t addr,
-                             uint8_t *txbuf, uint8_t len,
+                             uint8_t *txbuf, uint16_t len,
                              bool sendStop = true) = 0;
 
     // Set the speed in Hz. Typical values are
