+++ src/platform/msp432/i2c_msp432.h
@@ -35,9 +35,9 @@ public:
     virtual ~i2c_msp432();
 
     int16_t i2cRead (uint16_t addr, uint8_t *rxbuf,
-                     uint8_t len, bool sendStop=true) override;
+                     uint16_t len, bool sendStop=true) override;
     int16_t i2cWrite(uint16_t addr, uint8_t *txbuf,
-                     uint8_t len, bool sendStop=true) override;
+                     uint16_t len, bool sendStop=true) override;
 
     void setSpeed(uint32_t) override;
